Move variable definition to ty module.
