+++ sway-core/src/semantic_analysis/namespace/module.rs
@@ -2,7 +2,7 @@ use crate::{
     error::*,
     language::{parsed::*, ty, Visibility},
     semantic_analysis::{
-        ast_node::{TyAstNode, TyAstNodeContent, TyVariableDeclaration},
+        ast_node::{TyAstNode, TyAstNodeContent},
         declaration::VariableMutability,
         TypeCheckContext,
     },
@@ -343,7 +343,7 @@ impl Module {
                 }
                 // if this is a const, insert it into the local namespace directly
                 if let ty::TyDeclaration::VariableDeclaration(ref var_decl) = decl {
-                    let TyVariableDeclaration {
+                    let ty::TyVariableDeclaration {
                         mutability, name, ..
                     } = &**var_decl;
                     if mutability == &VariableMutability::ExportedConst {
